refactor: let card form UI own its props
