+++ src/pages/card-form/ui/CardForm.tsx
@@ -1,10 +1,38 @@
 import type * as React from "react";
 import { useId } from "react";
 
+import type { CardId } from "@/entities/card";
 import { Button } from "@/shared/ui/button";
 import { TagList } from "@/shared/ui/content";
 import { Form, FormItem, Tag, Textarea } from "@/shared/ui/forms";
-import type { CardFormProps } from "../model/useCardFormState";
+import type { Option } from "@/shared/ui/forms";
+
+interface CardFormTagField extends Option {
+  input: React.ComponentProps<typeof Tag>;
+}
+
+interface CardFormFields {
+  frontText: React.ComponentProps<typeof Textarea>;
+  backText: React.ComponentProps<typeof Textarea>;
+  tags: CardFormTagField[];
+}
+
+export interface CardFormProps {
+  cardInfo: {
+    uniqueKey: string;
+    id: CardId;
+    createdAt?: string;
+    lastSeenAt?: string;
+  };
+  fields: CardFormFields;
+  errors: {
+    frontText: string | undefined;
+    backText: string | undefined;
+  };
+  isSubmitting: boolean;
+  onCancel: () => void;
+  onSubmit: NonNullable<React.ComponentProps<typeof Form>["onSubmit"]>;
+}
 
 export const CardForm: React.FC<CardFormProps> = (props) => {
   const sectionHeadingIdPrefix = useId();
